Dynamically create Fields classes from functions
